runtime resource could be decimal

+++ server/controller/src/main/java/ai/starwhale/mlops/domain/runtime/RuntimeResource.java
@@ -29,12 +29,12 @@ import lombok.NoArgsConstructor;
 public class RuntimeResource {
 
     String type;
-    Integer num;
+    Float num;
 
     public RuntimeResource(String str) {
         String[] res = str.split("=");
         this.type = res[0];
-        this.num = Integer.valueOf(res[1]);
+        this.num = Float.valueOf(res[1]);
     }
 
 }
